moved static field reset function to its own class
