+++ Plugin/ContainerRegister.cs
@@ -15,6 +15,7 @@ using UniTASPlugin.Movie;
 using UniTASPlugin.Movie.EngineMethods;
 using UniTASPlugin.Patches.PatchProcessor;
 using UniTASPlugin.ReverseInvoker;
+using UniTASPlugin.StaticFieldStorage;
 using UniTASPlugin.Trackers.SceneIndexNameTracker;
 using UniTASPlugin.Trackers.SceneTracker;
 
@@ -25,6 +26,23 @@ public static class ContainerRegister
     public static Container Init()
     {
         var container = new Container(c =>
+        {
+            c.For<StaticFieldStorage.StaticFieldStorage>().Singleton();
+            c.For<IStaticFieldManipulator>().Use(x => x.GetInstance<StaticFieldStorage.StaticFieldStorage>());
+
+            // note: this code on initial load needs to run before destroying game objects
+            c.For<SceneIndexNameTracker>().Singleton();
+            c.For<ISceneIndexName>().Use(x => x.GetInstance<SceneIndexNameTracker>());
+            c.For<IPluginInitialLoad>().Use(x => x.GetInstance<SceneIndexNameTracker>());
+            c.For<IOnUpdate>().Use(x => x.GetInstance<SceneIndexNameTracker>());
+        });
+
+        return container;
+    }
+
+    public static void ConfigAfterInit(Container container)
+    {
+        container.Configure(c =>
         {
             c.Scan(scanner =>
             {
@@ -84,17 +102,9 @@ public static class ContainerRegister
             c.For<IObjectTracker>().Use(x => x.GetInstance<ObjectTracker>());
             c.For<IObjectInfo>().Use(x => x.GetInstance<ObjectTracker>());
 
-            // note: this code on initial load needs to run before destroying game objects
-            c.For<SceneIndexNameTracker>().Singleton();
-            c.For<ISceneIndexName>().Use(x => x.GetInstance<SceneIndexNameTracker>());
-            c.For<IPluginInitialLoad>().Use(x => x.GetInstance<SceneIndexNameTracker>());
-            c.For<IOnUpdate>().Use(x => x.GetInstance<SceneIndexNameTracker>());
-
             c.For<SceneTracker>().Singleton();
             c.For<ISceneTracker>().Use(x => x.GetInstance<SceneTracker>());
             c.For<ILoadedSceneInfo>().Use(x => x.GetInstance<SceneTracker>());
         });
-
-        return container;
     }
 }
