+++ Plugin/StaticFieldStorage/IStaticFieldManipulator.cs
@@ -1,0 +1,6 @@
+namespace UniTASPlugin.StaticFieldStorage;
+
+public interface IStaticFieldManipulator
+{
+    void ResetStaticFields();
+}
